Rename files according to rangle.io guidelines.
1. No uppercase letters in file names, this is not our convention and causes problems with cross-platform teams.
2. Only use `.tsx` for files that actually contain TSX code.

Connected to rangle/rangle-starter#23
