+++ src/components/counter.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import Button from './Button';
+import Button from './button';
 
 interface ICounterProps extends React.Props<any> {
   counter?: number;
