+++ src/components/navigator-item.tsx
