+++ src/utils/is-promise.ts
