+++ test/mocha.opts
@@ -1,4 +1,4 @@
 --compilers ts:ts-node/register
 --compilers tsx:ts-node/register
---require ./test/entry.tsx
-**/*.test.tsx
+--require ./test/entry.ts
+**/*.test.tsx **/*.test.ts
